Refactor controller tests to use new FakeInjectionController, FakeTypedParamsController, and FakeUnionTypesController; remove deprecated PHP71 controller references.
