+++ tests/framework/console/FakeInjectionController.php
@@ -6,27 +6,21 @@
  * @license https://www.yiiframework.com/license/
  */
 
-namespace yiiunit\framework\web;
+namespace yiiunit\framework\console;
 
 use yii\data\DataProviderInterface;
-use yii\web\Controller;
-use yii\web\Request;
-use yiiunit\framework\web\stubs\VendorImage;
-use yiiunit\framework\web\stubs\ModelBindingStub;
+use yiiunit\framework\console\stubs\DummyService;
+use yii\console\Controller;
+use yii\console\Request;
+use yiiunit\framework\base\Post;
 
-/**
- * @author Sam Mousa<sam@mousa.nl>
- * @since 2.0.36
- */
-class FakePhp71Controller extends Controller
+class FakeInjectionController extends Controller
 {
-    public $enableCsrfValidation = false;
-
     public function actionInjection(
         $before,
         Request $request,
         $between,
-        VendorImage $vendorImage,
+        DummyService $dummyService,
         ?Post $post,
         $after
     ) {
@@ -37,10 +31,6 @@ class FakePhp71Controller extends Controller
     }
 
     public function actionModuleServiceInjection(DataProviderInterface $dataProvider): void
-    {
-    }
-
-    public function actionModelBindingInjection(ModelBindingStub $model): void
     {
     }
 }
