+++ tests/framework/di/stubs/StaticMethodsWithIntersectionTypes.php
@@ -2,7 +2,6 @@
 
 namespace yiiunit\framework\di\stubs;
 
-// Syntax valid only for PHP 8.1+
 class StaticMethodsWithIntersectionTypes
 {
     public static function withQuxInterfaceAndQuxAnotherIntersection(QuxInterface & QuxAnother $Qux): void
