+++ tests/framework/di/stubs/StaticMethodsWithUnionTypes.php
@@ -2,7 +2,6 @@
 
 namespace yiiunit\framework\di\stubs;
 
-// Syntax valid only for PHP 8.0+
 class StaticMethodsWithUnionTypes
 {
     public static function withBetaUnion(string | Beta $beta): void
