+++ tests/framework/validators/EachValidatorTest.php
@@ -218,8 +218,6 @@ class EachValidatorTest extends TestCase
      * of different type during validation.
      * (ie: public array $dummy; where $dummy is array of booleans,
      * validator will try to assign these booleans one by one to $dummy)
-     *
-     * @requires PHP >= 7.4
      */
     public function testTypedProperties(): void
     {
